Move descriptor.go into metric.go
